Drop nodejs 6 for next (Fastify 3) (#1713)
* update requirements to Node.js 8 LTS or later; update docs

add TODO for code updates/cleanup (check if do here or if in the other issue/PR)

Update README.md


Update docs/Benchmarking.md


Update package.json


revert version

update getHeader to use hasHeader (available since Node.js 7.7.x) and remove fallback implementation

try yarn builds for Node.js 8 aligned with requirements in 'package.json', to check if it works with version range

align even npm azure pipelines to those of yarn (Node.js 8.x requirements)

update code to Node.js 8 as minimum version (remove TODO etc)

update code to Node.js 8 as minimum version (remove TODO etc)

update code to Node.js 8 as minimum version (remove TODO etc)

move getHeader implementation directly in Reply.prototype.getHeader

drop Node.js 11 from doc and from CI on Azure (yarn and npm)

(try to) simplify info on HTTP2 support

CI on Azure, specify version range in a way more aligned to npm syntax

* restore 'hooks-async.js' and set the '.test.js' extension to run it like others; remove its tests from 'hooks.test.js'

* remove direct reference to main fastify source from https tests, not really needed
