+++ README.md
@@ -30,6 +30,10 @@ How can you efficiently handle the resources of your server, knowing that you ar
 
 Enter Fastify. Fastify is a web framework highly focused on providing the best developer experience with the least overhead and a powerful plugin architecture. It is inspired by Hapi and Express and as far as we know, it is one of the fastest web frameworks in town.
 
+### Requirements
+
+Node.js 8 LTS (8.16.0) or later; recommended 10 LTS (10.16.0) or later.
+
 ### Install
 
 Install with npm:
@@ -105,9 +109,10 @@ fastify generate
 
 For more information, see the [Fastify CLI documentation](https://github.com/fastify/fastify-cli).
 
-### Fastify v1.x
+### Fastify v1.x and v2.x
 
 Code for Fastify's **v1.x** is in [Branch 1.x](https://github.com/fastify/fastify/tree/1.x), so all Fastify 1.x related changes should be based on **`branch 1.x`**.
+In a similar way, all Fastify 2.x related changes should be based on **`branch 2.x`**.
 
 > ## Note
 > `.listen` binds to the local host, `localhost`, interface by default (`127.0.0.1` or `::1`, depending on the operating system configuration). If you are running Fastify in a container (Docker, [GCP](https://cloud.google.com/), etc.), you may need to bind to `0.0.0.0`. Be careful when deciding to listen on all interfaces; it comes with inherent [security risks](https://web.archive.org/web/20170711105010/https://snyk.io/blog/mongodb-hack-and-secure-defaults/).
