+++ docs/HTTP2.md
@@ -2,9 +2,9 @@
 
 ## HTTP2
 
-_Fastify_ offers **experimental support** for HTTP2 starting from Node
-8.8.0, which includes HTTP2 without a flag. _Fastify_ supports HTTP2
-both over HTTPS or over plaintext. Note that HTTP2 is available only for node versions >= `8.8.1`.
+_Fastify_ offers **experimental support** for HTTP2 starting from 
+Node 8 LTS, which includes HTTP2 without a flag; HTTP2 is supported 
+both over HTTPS or over plaintext.
 
 Currently none of the HTTP2-specific APIs are available through
 _Fastify_, but Node's `req` and `res` can be access through our
