+++ lib/reply.js
@@ -50,8 +50,6 @@ const {
   }
 } = require('./errors')
 
-var getHeader
-
 function Reply (res, context, request, log) {
   this.res = res
   this.context = context
@@ -115,7 +113,7 @@ Reply.prototype.send = function (payload) {
     return this
   }
 
-  var contentType = getHeader(this, 'content-type')
+  var contentType = this.getHeader('content-type')
   var hasContentType = contentType !== undefined
 
   if (payload !== null) {
@@ -151,7 +149,13 @@ Reply.prototype.send = function (payload) {
 }
 
 Reply.prototype.getHeader = function (key) {
-  return getHeader(this, key)
+  key = key.toLowerCase()
+  var res = this.res
+  var value = this[kReplyHeaders][key]
+  if (value === undefined && res.hasHeader(key)) {
+    value = res.getHeader(key)
+  }
+  return value
 }
 
 Reply.prototype.hasHeader = function (key) {
@@ -566,37 +570,6 @@ function notFound (reply) {
 }
 
 function noop () {}
-
-function getHeaderProper (reply, key) {
-  key = key.toLowerCase()
-  var res = reply.res
-  var value = reply[kReplyHeaders][key]
-  if (value === undefined && res.hasHeader(key)) {
-    value = res.getHeader(key)
-  }
-  return value
-}
-
-function getHeaderFallback (reply, key) {
-  key = key.toLowerCase()
-  var res = reply.res
-  var value = reply[kReplyHeaders][key]
-  if (value === undefined) {
-    value = res.getHeader(key)
-  }
-  return value
-}
-
-// ponyfill for hasHeader. It has been introduced into Node 7.7,
-// so it's ok to use it in 8+
-{
-  const v = process.version.match(/v(\d+)/)[1]
-  if (Number(v) > 7) {
-    getHeader = getHeaderProper
-  } else {
-    getHeader = getHeaderFallback
-  }
-}
 
 module.exports = Reply
 module.exports.buildReply = buildReply
