+++ test/fluent-schema.test.js
@@ -1,11 +1,121 @@
 'use strict'
 
 const t = require('tap')
-const semver = require('semver')
+const test = t.test
+const Fastify = require('..')
+const S = require('fluent-schema')
 
-if (semver.gt(process.versions.node, '8.0.0')) {
-  require('./fluent-schema')(t)
-} else {
-  t.pass('Skip because Node version < 8')
-  t.end()
-}
+test('fluent-schema generate a valid JSON Schema in "$ref-way"', t => {
+  t.plan(1)
+
+  const fastify = new Fastify()
+
+  const addressSchema = S.object()
+    .id('#address')
+    .prop('line1').required()
+    .prop('line2')
+    .prop('country').required()
+    .prop('city').required()
+    .prop('zipcode').required()
+    .valueOf()
+
+  const commonSchemas = S.object()
+    .id('https://fastify/demo')
+    .definition('addressSchema', addressSchema)
+    .valueOf()
+
+  fastify.addSchema(commonSchemas)
+
+  const bodyJsonSchema = S.object()
+    .prop('residence', S.ref('https://fastify/demo#address')).required()
+    .prop('office', S.ref('https://fastify/demo#/definitions/addressSchema')).required()
+    .valueOf()
+
+  const schema = { body: bodyJsonSchema }
+  fastify.post('/the/url', { schema }, () => { })
+
+  fastify.ready(t.error)
+})
+
+test('fluent-schema generate a valid JSON Schema in "replace-way"', t => {
+  t.plan(1)
+
+  const fastify = new Fastify()
+
+  const sharedAddressSchema = {
+    $id: 'sharedAddress',
+    type: 'object',
+    required: ['line1', 'country', 'city', 'zipcode'],
+    properties: {
+      line1: { type: 'string' },
+      line2: { type: 'string' },
+      country: { type: 'string' },
+      city: { type: 'string' },
+      zipcode: { type: 'string' }
+    }
+  }
+
+  fastify.addSchema(sharedAddressSchema)
+
+  const bodyJsonSchema = {
+    type: 'object',
+    properties: {
+      vacation: 'sharedAddress#'
+    }
+  }
+  const schema = { body: bodyJsonSchema }
+
+  fastify.post('/the/url', { schema }, () => { })
+
+  fastify.ready(t.error)
+})
+
+test('fluent-schema mix-up of "$ref-way" and "replace-way"', t => {
+  t.plan(1)
+
+  const fastify = new Fastify()
+
+  const addressSchema = S.object()
+    .id('#address')
+    .prop('line1').required()
+    .prop('line2')
+    .prop('country').required()
+    .prop('city').required()
+    .prop('zipcode').required()
+    .valueOf()
+
+  const commonSchemas = S.object()
+    .id('https://fastify/demo')
+    .definition('addressSchema', addressSchema)
+    .valueOf()
+
+  const sharedAddressSchema = {
+    $id: 'sharedAddress',
+    type: 'object',
+    required: ['line1', 'country', 'city', 'zipcode'],
+    properties: {
+      line1: { type: 'string' },
+      line2: { type: 'string' },
+      country: { type: 'string' },
+      city: { type: 'string' },
+      zipcode: { type: 'string' }
+    }
+  }
+
+  fastify.addSchema(commonSchemas)
+  fastify.addSchema(sharedAddressSchema)
+
+  const bodyJsonSchema = S.object()
+    .prop('residence', S.ref('https://fastify/demo#address')).required()
+    .prop('office', S.ref('https://fastify/demo#/definitions/addressSchema')).required()
+    .valueOf()
+
+  // add the key with the string value to use shared schema in "replace-way"
+  bodyJsonSchema.properties.vacation = 'sharedAddress#'
+
+  const schema = { body: bodyJsonSchema }
+
+  fastify.post('/the/url', { schema }, () => { })
+
+  fastify.ready(t.error)
+})
