+++ test/reply-error.test.js
@@ -3,7 +3,6 @@
 const t = require('tap')
 const test = t.test
 const net = require('net')
-const semver = require('semver')
 const Fastify = require('..')
 const statusCodes = require('http').STATUS_CODES
 
@@ -101,34 +100,32 @@ test('onRequest hook error handling with external done', t => {
   })
 })
 
-if (semver.gt(process.versions.node, '6.0.0')) {
-  test('Should reply 400 on client error', t => {
-    t.plan(2)
+test('Should reply 400 on client error', t => {
+  t.plan(2)
 
-    const fastify = Fastify()
-    fastify.listen(0, err => {
-      t.error(err)
+  const fastify = Fastify()
+  fastify.listen(0, err => {
+    t.error(err)
 
-      const client = net.connect(fastify.server.address().port)
-      client.end('oooops!')
+    const client = net.connect(fastify.server.address().port)
+    client.end('oooops!')
 
-      var chunks = ''
-      client.on('data', chunk => {
-        chunks += chunk
+    var chunks = ''
+    client.on('data', chunk => {
+      chunks += chunk
+    })
+
+    client.once('end', () => {
+      const body = JSON.stringify({
+        error: 'Bad Request',
+        message: 'Client Error',
+        statusCode: 400
       })
-
-      client.once('end', () => {
-        const body = JSON.stringify({
-          error: 'Bad Request',
-          message: 'Client Error',
-          statusCode: 400
-        })
-        t.equal(`HTTP/1.1 400 Bad Request\r\nContent-Length: ${body.length}\r\nContent-Type: application/json\r\n\r\n${body}`, chunks)
-        fastify.close()
-      })
+      t.equal(`HTTP/1.1 400 Bad Request\r\nContent-Length: ${body.length}\r\nContent-Type: application/json\r\n\r\n${body}`, chunks)
+      fastify.close()
     })
   })
-}
+})
 
 test('Error instance sets HTTP status code', t => {
   t.plan(3)
